Moved etos into their individual files
