+++ framework/src/Volo.Abp.Auditing/Volo/Abp/Auditing/AuditLogInfoEto.cs
@@ -57,46 +57,3 @@ public class AuditLogInfoEto : EtoBase, IHasExtraProperties
     public List<string> Comments { get; set; }
     public ExtraPropertyDictionary ExtraProperties { get; set; }
 }
-
-[Serializable]
-public class AuditLogActionInfoEto : IHasExtraProperties
-{
-    public string ServiceName { get; set; }
-
-    public string MethodName { get; set; }
-
-    public string Parameters { get; set; }
-
-    public DateTime ExecutionTime { get; set; }
-
-    public int ExecutionDuration { get; set; }
-    public ExtraPropertyDictionary ExtraProperties { get; set; }
-}
-
-[Serializable]
-public class EntityChangeInfoEto : IHasExtraProperties
-{
-    public DateTime ChangeTime { get; set; }
-
-    public EntityChangeType ChangeType { get; set; }
-    public Guid? EntityTenantId { get; set; }
-
-    public string EntityId { get; set; }
-
-    public string EntityTypeFullName { get; set; }
-
-    public List<EntityPropertyChangeInfoEto> PropertyChanges { get; set; }
-    public ExtraPropertyDictionary ExtraProperties { get; set; }
-}
-
-[Serializable]
-public class EntityPropertyChangeInfoEto
-{
-    public string NewValue { get; set; }
-
-    public string OriginalValue { get; set; }
-
-    public string PropertyName { get; set; }
-
-    public string PropertyTypeFullName { get; set; }
-}
