+++ framework/src/Volo.Abp.Auditing/Volo/Abp/Auditing/EntityPropertyChangeInfoEto.cs
@@ -1,0 +1,15 @@
+using System;
+
+namespace Volo.Abp.Auditing;
+
+[Serializable]
+public class EntityPropertyChangeInfoEto
+{
+    public string NewValue { get; set; }
+
+    public string OriginalValue { get; set; }
+
+    public string PropertyName { get; set; }
+
+    public string PropertyTypeFullName { get; set; }
+}
